feat: декомпозиция поиска
