+++ frontend/src/widgets/index.ts
@@ -1,2 +1,3 @@
 export { Footer } from './footer/Footer';
 export { Header } from './header/Header';
+export { SearchAdBlock } from './advertisementSearch/SearchAdBlock.tsx';
